delete_confirmation

+++ admin/includes/users_viewAll.php
@@ -50,7 +50,7 @@
             echo "<td><a href='users.php?change_to_admin={$user_id}'>Make admin</a></td>";
             echo "<td><a href='users.php?change_to_subsc={$user_id}'>Make subscriber</a></td>";
             echo "<td><a href='users.php?source=edit_user&user_id={$user_id}'>Edit</a></td>";
-            echo "<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this user?'); \" href='users.php?delete={$user_id}'>Delete</a></td>";
             echo "</tr>";
         }
